+++ src/components/ToggleButton.vue
@@ -1,0 +1,32 @@
+<template>
+  <div>
+    <v-btn
+      icon
+      @click.stop="toggleButton"
+    >
+      <v-icon>mdi-chevron-left</v-icon>
+    </v-btn>
+  </div>
+</template>
+
+<script lang="ts">
+import { defineComponent } from '@nuxtjs/composition-api'
+export default defineComponent({
+  props: {
+    miniProp: {
+      type: Boolean,
+      default: true
+    }
+  },
+
+  setup (props, context) {
+    const toggleButton = () => {
+      context.emit('change-button', !props.miniProp)
+    }
+
+    return {
+      toggleButton
+    }
+  }
+})
+</script>
