+++ src/pages/index.vue
@@ -19,10 +19,10 @@
 
 <script lang="ts">
 import { defineComponent } from '@vue/composition-api'
-import Home from '@/components/Home.vue'
-import Side from '@/components/Side.vue'
-import HeaderContent from '@/components/Header.vue'
-import FooterContent from '@/components/Footer.vue'
+import Side from '../components/Side.vue'
+import Home from '../components/Home.vue'
+import HeaderContent from '../components/Header.vue'
+import FooterContent from '../components/Footer.vue'
 
 export default defineComponent({
   components: {
