+++ src/plugins/amplify.js
@@ -1,7 +1,7 @@
 import Vue from 'vue'
 import { Amplify } from 'aws-amplify'
 import '@aws-amplify/ui-vue'
-import awsExports from '../src/aws-exports'
+import awsExports from '@/aws-exports'
 
 Amplify.configure(awsExports)
 Vue.use(Amplify)
