+++ src/store/post/type.ts
